Use real Maven dependencies on sister projects, fixes #18
This change updates the build configuration to properly capture the
dependency on two sister projects within the Gradle build process
itself. This also means that the build process is much simpler for
people who wish to contribute changes to this project; they no longer
need to check out or build the sister projects, as that is taken care
of by the Maven infrastructure.

I have also verified that this change fixes the generated `pom.xml` so
that client projects will automatically know they need to have
RSyntaxTextarea and AutoComplete on the class path. Beat Link Trigger
no longer crashes when I depend on this snapshot version, as it did
when I depended on 3.0.3.

To test against a local (not yet published) snapshot of one of the
sister projects, all you need to do is configure this project to
depend on the snapshot version, and then run `gradlew
publishToMavenLocal` in the sister project you want to test against.

Before releasing this, version 3.0.8 of RSyntaxTextArea must be
released, and this `build.gradle` needs to be updated to depend on the
released version.